Auto-format code with Prettier
Prettier has been set up as the auto-formatter for this project.
Enforcing a common style avoids unnecessary friction in pull request
reviews, and makes it easier to contribute to the code base.

Prettier has been added to the pre-commit hooks of the repository to
ensure that every commit is correctly formatted. This is verified by a
GitHub Action that has to pass before code can be merged into the main
branch.