Set up TypeScript (#4)
The action is implemented using TypeScript, and then compiled to
Javascript. After compiling the code with the TypeScript compiler `tsc`,
the action and its dependencies are packaged into a single file using
`ncc`.

This is all done in a pre-commit hook so that the repository always
contains the latest version of the action. A GitHub Action checks and
enforces this.