Prototype action
A first version of the action has been implemented. The implementation
is inspired by `actions/setup-node` [^1] and `buf-setup-action` [^2].

The action checks if butler has previously been cached, and if so
restores the binary from the cache. If it hasn't been set up before, it
is downloaded from itch.io's package repository.

[^1]: https://github.com/actions/setup-node
[^2]: https://github.com/bufbuild/buf-setup-action