Update domain for itch's CDN
itch.io recently experienced issues with the domain they used for their content
delivery network, `itch.ovh`. They created `itch.zone` to replace the now
deprecated domain, and #421 updated the script to use it. This commit recompiles
the action so that the new domain will be used from now on.